requirements.txt.txt
# Core dependencies
streamlit
google-auth
google-auth-oauthlib
google-api-python-client

# LangChain & Groq
langchain
langchain-groq

# For working with OAuth credentials and HTTP
requests

# Optional: Safe file handling & others
tqdm

# Exact version pins (optional for stability)
# streamlit==1.35.0
# google-auth==2.29.0
# google-auth-oauthlib==1.2.0
# google-api-python-client==2.127.0
# langchain==0.2.0
# langchain-groq==0.1.3
